refactoring: 파일/폴더 정리, pages/Detail.jsx 삭제
