+++ src/App.jsx
@@ -1,8 +1,8 @@
 import { BrowserRouter, Route, Routes } from "react-router-dom";
 import GlobalStyle from "../style/GlobalStyle";
 import HomeLayout from "./components/layouts/HomeLayout";
-import SignIn from "./pages/SignIn";
-import SignUp from "./pages/SignUp";
+import SignIn from "./pages/Sign/SignIn";
+import SignUp from "./pages/Sign/SignUp";
 
 const App = () => {
   return (
