--- src/pages/Detail.jsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const Detail = () => {
-  return <div>2222222테스트</div>;
-};
-
-export default Detail;
